[Refactoring] : HUD에서 리스폰 패널 제거
UIManager로 기능 넘김
